Revert "msm8916-common overlay: kill pinner service"
Removing systemUI from pinner services causes a 4 to 6 delay in
bringing up the home screen.  This is seen when loading or
playing a game for about 20 minutes and then either hitting
the recents or home button.

By reverting this commit, pressing recents or home button
brings up the systemUI right away.

This reverts commit c0e470a42f09ce9f74564fb74e805fbd54dae2b2.