msm8916-common: remove audio hal entries because it's not common to all msm8916 tablets
The SM-T350 uses a proprietary sound driver for the amplifier and only works with audio 2.0.
This has been the case since the beginning and not specific to Android 12.

So we need to put this in the specific device tree and not in msm8916-common.

In addition, Android 12 drops audio 2.0, but it can be reverted using commits by Tarkzim as per

https://github.com/TARKZiM/android_hardware_interfaces/commit/692199ff35ac62d305d6216f13d8d69484efe0a0
https://github.com/TARKZiM/android_frameworks_av/commit/a559c9ecb9804002c625037c37481bce87a972f9